pyproject.toml
[project]
name = "concordia"
version = "0.1.0"
description = "Multi-user prompt party with dedupe and Claude Code execution"
readme = "README.md"
requires-python = ">=3.9"
authors = [{name = "Concordia Contributors"}]
license = {text = "MIT"}

dependencies = [
  "websockets>=12.0",
  "python-dotenv>=1.0.0",
  "requests>=2.31.0",
]

[project.scripts]
concordia = "concordia.cli:main"
concordia_host = "concordia.host_cli:main"
concordia_client = "concordia.client_cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
